Cleanup stories
